+++ .env
@@ -1,0 +1,3 @@
+LOGIN=technopark26
+PASSWORD=testQA1
+BROWSER=FIREFOX
