+++ run_tests.py
@@ -1,13 +1,13 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 
 import sys
 import unittest
-from tests.example_test import ExampleTest
+from tests.album_test import AlbumTest
 
 
 if __name__ == '__main__':
     suite = unittest.TestSuite((
-        unittest.makeSuite(ExampleTest),
+        unittest.makeSuite(AlbumTest),
     ))
     result = unittest.TextTestRunner().run(suite)
     sys.exit(not result.wasSuccessful())
