+++ tests/pages/page.py
@@ -1,0 +1,13 @@
+import urllib.parse
+
+class Page(object):
+    BASE_URL = 'http://tech-mail.ru/'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+
+    def open(self):
+        url = urllib.parse.urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
